+++ ajax/commonitilobjectiteminformation.php
@@ -32,15 +32,50 @@
  */
 
 // Direct access to file
-if (strpos($_SERVER['PHP_SELF'], "ticketiteminformation.php")) {
-    $AJAX_INCLUDE = 1;
-    include('../inc/includes.php');
-    header("Content-Type: text/html; charset=UTF-8");
-    Html::header_nocache();
+
+
+if (isset($_POST["my_items"]) && !empty($_POST["my_items"])) {
+    $splitter = explode("_", $_POST["my_items"]);
+    if (count($splitter) == 2) {
+        $_POST["itemtype"] = $splitter[0];
+        $_POST["items_id"] = $splitter[1];
+    }
 }
 
-Session::checkLoginUser();
+if (
+    isset($_POST['itemtype'])
+    && isset($_POST['items_id']) && ($_POST['items_id'] > 0)
+) {
+   // Security
+    if (!class_exists($_POST['itemtype'])) {
+        exit();
+    }
 
-$obj_class = 'Ticket';
-$item_obj_class = 'Item_Ticket';
-include('./commonitilobjectiteminformation.php');
+    $days   = 3;
+    $obj = new $obj_class();
+    $data   = $obj->getActiveOrSolvedLastDaysForItem(
+        $_POST['itemtype'],
+        $_POST['items_id'],
+        $days
+    );
+
+    $nb = count($data);
+    $badge_helper = sprintf(
+        _n(
+            '%s ticket in progress or recently solved on this item.',
+            '%s tickets in progress or recently solved on this item.',
+            $nb
+        ),
+        $nb
+    );
+    echo "<span class='badge badge-secondary' title='$badge_helper'>$nb</span>";
+
+    if ($nb) {
+        $content = '';
+        foreach ($data as $title) {
+            $content .= $title . '<br>';
+        }
+        echo '&nbsp;';
+        Html::showToolTip($content);
+    }
+}
