+++ ajax/item_ticket.php
@@ -31,16 +31,13 @@
  * ---------------------------------------------------------------------
  */
 
-// Direct access to file
-if (strpos($_SERVER['PHP_SELF'], "ticketiteminformation.php")) {
-    $AJAX_INCLUDE = 1;
-    include('../inc/includes.php');
-    header("Content-Type: text/html; charset=UTF-8");
-    Html::header_nocache();
-}
+include('../inc/includes.php');
+
+header("Content-Type: text/html; charset=UTF-8");
+Html::header_nocache();
 
 Session::checkLoginUser();
 
-$obj_class = 'Ticket';
-$item_obj_class = 'Item_Ticket';
-include('./commonitilobjectiteminformation.php');
+$obj = 'Ticket';
+$item_obj = 'Item_Ticket';
+include('./object_item.php');
